requirements.txt
Django==2.0.8
djangorestframework==3.8.2
pkg-resources==0.0.0
pytz==2018.5
